Started styling Login/Signup, added signup functionality
Added a bit of styling to login and signup pages. Signup page can save new users to database and redirect to a placeholder messaging page. Login page can also check if login data is in database and redirects to the same messaging page. More styling,edits to be done in future.