"Working" state
Added socketIo functionality, ability to search for users, and ability to retrieve stored messages and display them. Made some adjustments in routes to do less fetches and conversations are no longer saved on users.

This 'works'. There is a huge issue with performance currently. This is likely due to the number of rerenders the application goes through. Prop drilling probably a major contributor. May need to consolidate some components as well such as the search component which is only used once.

Needs more polishing.