Refactored to improve performance
Mainly refactored the messaging page structure so that there is a new top level component whose sole job is to connect to the server (socket) with the right roomId if the roomId has been retrieved.
Message text field no longer rerenders the entire message panel component and is in a component of its own.
Added conditional logic to the conversation cards to account for cases such as when there are no current conversations to retrieve, etc.