Bug fixes and some styling
Was a bug in which receiving a message would exponentially increase the time it took for the next one to arrive.

Turns out it was an oversight where socket event listeners were never turned off and react rerenders exacerbated the problem.

Also added a little more styling and refactored a bit to prevent some non essential rerenders.